+++ AzureDevOpsBackup/Class/Globals.cs
@@ -1,5 +1,6 @@
-﻿using System.Net.Mail;
-using System.Security.Policy;
+﻿using System;
+using System.Net.Mail;
+using static AzureDevOpsBackup.Class.FileLogger;
 
 namespace AzureDevOpsBackup.Class
 {
@@ -25,5 +26,18 @@ namespace AzureDevOpsBackup.Class
         public static int _currentBackupsInBackupFolderCount;
         public static int _oldLogFilesToDeleteCount;
         public static bool _oldLogfilesToDelete;
+
+        public static void ApplicationStartMessage()
+        {
+            // Log start of program
+            Message($"Welcome to {Globals.AppName}, v." + Globals._vData + " by " + Globals._companyName, EventType.Information, 1000);
+            Console.WriteLine($"\nWelcome to {Globals.AppName}, v." + Globals._vData + " by " + Globals._companyName + "\n");
+        }
+        public static void ApplicationEndMessage()
+        {
+            // Log end of program
+            Message($"End of application - {Globals.AppName}, v." + Globals._vData, EventType.Information, 1000);
+            Console.WriteLine($"\nEnd of application - {Globals.AppName}, v. {Globals._vData}\n");
+        }
     }
 }