+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ### Add
 - Added Class to secure the token for authentication to the Azure DevOps API form an encrypted .bin file based in hardware id for the hardware the .bin file is generated on
 
+### Changed
+- Some changes to help text
+
 ## [1.0.5.3] - 29-3-2023
 
 ### Add